requirements.txt
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
six==1.16.0
TgCrypto==1.2.5
